+++ app-api/src/main/kotlin/site/techmoa/app/auth/v1/AuthController.kt
@@ -4,7 +4,7 @@ import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestParam
 import org.springframework.web.bind.annotation.RestController
 import site.techmoa.app.auth.v1.response.LoginResponse
-import site.techmoa.app.core.member.KakaoLoginOrSignupFacade
+import site.techmoa.app.core.auth.kakao.KakaoLoginOrSignupFacade
 import site.techmoa.app.core.response.ApiResponse
 
 @RestController
