+++ app-core/src/main/kotlin/site/techmoa/app/core/auth/kakao/port/KakaoOidcClient.kt
@@ -1,0 +1,8 @@
+package site.techmoa.app.core.auth.kakao.port
+
+import site.techmoa.app.core.auth.kakao.dto.KakaoToken
+
+interface KakaoOidcClient {
+
+    fun getToken(code: String): KakaoToken
+}
