--- app-core/src/main/kotlin/site/techmoa/app/core/member/KakaoOidcClient.kt
@@ -1,6 +1,0 @@
-package site.techmoa.app.core.member
-
-interface KakaoOidcClient {
-
-    fun getToken(code: String): KakaoToken
-}
