--- app-core/src/main/kotlin/site/techmoa/app/core/member/KakaoOidcExchanger.kt
@@ -1,12 +1,0 @@
-package site.techmoa.app.core.member
-
-import org.springframework.stereotype.Component
-
-@Component
-class KakaoOidcExchanger(
-    private val kakaoOidcClient: KakaoOidcClient
-) {
-    fun retrieve(code: String): KakaoToken {
-        return kakaoOidcClient.getToken(code)
-    }
-}
